backend/requirements-pip.txt
djangorestframework==3.12.4
gunicorn==20.0.4
persiantools==1.4.1
openpyxl==3.0.3
django-nose==1.4.7
coverage==5.3
django-cors-headers==3.7.0
drf-recaptcha==2.0.4
drf-spectacular==0.17.2
celery==5.2.2
django-celery-beat==2.2.1
google-auth==2.6.0
requests==2.27.1
django-test-migrations==1.2.0
